backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-pptx==0.6.21
openai==0.28.1
anthropic==0.3.11
google-generativeai==0.3.0
Pillow==10.0.0
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0